refactor: 💡 move logic accordingly
