+++ app/helpers/episodes_helper.rb
@@ -9,6 +9,14 @@ module EpisodesHelper
     episode.try(:preview?) ? preview_episode_path(episode) : episode_path(episode)
   end
 
+  def auto_link_url(text)
+    auto_link(text, html: { target: '_blank' })
+  end
+
+  def simple_format_with_link_new(text)
+    simple_format(text, {}, { sanitize_options: { attributes: %w[target href] } })
+  end
+
   def markdown_link(text)
     if text.match(/\[.*\]\(.*\)/)
       text.gsub(/\[(.*?)\]\((.*?)\)/, link_to('\1', '\2', target: '_blank', rel: 'noopener'))
